Extract release-safe headless runner (raxol p)
The headless surface lived only in Mix.Tasks.Raxol.P, and a Burrito
release has no Mix -- so the packaged CLI binary (the T-Bench artifact)
had no way to run it. The whole runner moves to Raxol.Agent.P (no Mix
calls, returns the exit code, usage errors via a contained throw); the
mix task becomes a boot shim that delegates; the raxol CLI dispatches
'raxol p' / 'raxol -p' straight to it, so the same code path serves dev
and the release binary. Also underscores an unused param in CLI.New
that broke --warnings-as-errors for the package.
