+++ packages/raxol_agent/lib/mix/tasks/raxol.p.ex
@@ -63,56 +63,12 @@ defmodule Mix.Tasks.Raxol.P do
 
   use Mix.Task
 
-  alias Raxol.Agent.BenchmarkProfile
-  alias Raxol.Agent.Contract
-  alias Raxol.Agent.SessionStreamer
-  alias Raxol.Agent.SignalTrap
-  alias Raxol.Agent.Trajectory
-
-  @default_timeout_s 180
-
-  @switches [
-    backend: :string,
-    # `--harness` is a deprecated alias for `--backend`.
-    harness: :string,
-    model: :string,
-    base_url: :string,
-    system: :string,
-    timeout: :integer,
-    write: :boolean,
-    tools: :boolean
-  ]
-
   @impl Mix.Task
   def run(argv) do
-    {opts, args, invalid} = OptionParser.parse(argv, strict: @switches)
-
-    prompt = Enum.join(args, " ") |> String.trim()
-
-    cond do
-      invalid != [] ->
-        usage_error("unknown options: #{inspect(invalid)}")
-
-      prompt == "" ->
-        usage_error("no prompt given. Usage: mix raxol.p [options] \"prompt\"")
-
-      true ->
-        run_prompt(prompt, opts)
-    end
-  end
-
-  defp usage_error(message) do
-    IO.puts(:stderr, "raxol.p: #{message}")
-    exit({:shutdown, 64})
-  end
-
-  defp run_prompt(prompt, opts) do
-    # Agent environment only — no terminal driver, no UI. stdout belongs to
-    # the answer: boot without recompiling (bin/raxol precompiles silently)
-    # and keep Logger quiet below :error.
-    System.put_env("RAXOL_SKIP_TERMINAL_INIT", "true")
-    Logger.configure(level: :error)
-
+    # Boot without recompiling (bin/raxol precompiles silently) so stdout
+    # stays clean for the answer, then hand off to the shared runner --
+    # `Raxol.Agent.P` contains no Mix calls, so the same code path serves
+    # the Burrito-packaged `raxol p` where Mix does not exist.
     Mix.Task.run("app.start", [
       "--no-compile",
       "--no-deps-check",
@@ -120,280 +76,6 @@ defmodule Mix.Tasks.Raxol.P do
       "--no-elixir-version-check"
     ])
 
-    profile =
-      case BenchmarkProfile.from_env() do
-        {:ok, profile} -> profile
-        {:error, message} -> usage_error(message)
-      end
-
-    # Claim SIGTERM unconditionally: the BEAM default turns it into a clean
-    # exit 0, which a harness reads as success. We flush and exit 143.
-    SignalTrap.install(self())
-
-    ensure_streamer!()
-
-    session_id = "cli-#{System.unique_integer([:positive])}"
-    :ok = SessionStreamer.subscribe(session_id)
-
-    stream_opts = build_stream_opts(prompt, opts, profile)
-    use_tools = Keyword.get(opts, :tools, true)
-
-    runner =
-      Task.async(fn ->
-        stream =
-          if use_tools do
-            Raxol.Agent.Stream.react(prompt, stream_opts)
-          else
-            Raxol.Agent.Stream.run(prompt, stream_opts)
-          end
-
-        Contract.pump(session_id, stream, prompt: prompt)
-      end)
-
-    timeout_ms = Keyword.get(opts, :timeout, @default_timeout_s) * 1_000
-
-    state = %{
-      wrote_stdout: false,
-      profile: profile,
-      prompt: prompt,
-      backend: Keyword.get(stream_opts, :executor).backend,
-      turns: 0,
-      usage: %{input_tokens: 0, output_tokens: 0},
-      events: if(profile.trajectory_path, do: [], else: nil)
-    }
-
-    status = consume(session_id, runner, timeout_ms, state)
-    exit({:shutdown, status})
+    exit({:shutdown, Raxol.Agent.P.run(argv)})
   end
-
-  # SessionStreamer is not in any package supervision tree yet (the CLI is
-  # its first live consumer); start it idempotently.
-  defp ensure_streamer! do
-    case SessionStreamer.start_link([]) do
-      {:ok, _pid} ->
-        :ok
-
-      {:error, {:already_started, _pid}} ->
-        :ok
-
-      {:error, reason} ->
-        raise "cannot start SessionStreamer: #{inspect(reason)}"
-    end
-  end
-
-  defp build_stream_opts(_prompt, opts, profile) do
-    # raxol.p reserves stderr for the JSONL event stream, so pass prog: nil to
-    # suppress the plain-text deprecation notice that would corrupt it.
-    # CLI flags win over the env profile; the profile wins over defaults.
-    opts = apply_profile_defaults(opts, profile)
-
-    backend =
-      case Raxol.Agent.Backend.Cli.resolve(opts, nil) do
-        {:ok, backend} -> backend
-        {:error, message} -> usage_error(message)
-      end
-
-    executor_attrs =
-      [backend: backend]
-      |> maybe_put(:model, Keyword.get(opts, :model))
-
-    executor = Raxol.Agent.ExecutorConfig.new(executor_attrs)
-
-    backend_opts =
-      []
-      |> maybe_put(:base_url, Keyword.get(opts, :base_url))
-
-    system =
-      Keyword.get(
-        opts,
-        :system,
-        "You are a helpful assistant running in a terminal at the user's " <>
-          "current working directory. Use the available tools to inspect " <>
-          "files when the question is about them. Be concise."
-      )
-
-    write? = Keyword.get(opts, :write, false)
-
-    [
-      executor: executor,
-      backend_opts: backend_opts,
-      system_prompt: system,
-      actions: actions_for(write?, profile)
-    ] ++ context_for(write?, profile)
-  end
-
-  defp apply_profile_defaults(opts, %BenchmarkProfile{} = profile) do
-    opts
-    |> maybe_default(:backend, profile.backend && Atom.to_string(profile.backend))
-    |> maybe_default(:model, profile.model)
-  end
-
-  defp maybe_default(opts, _key, nil), do: opts
-
-  defp maybe_default(opts, key, value) do
-    if Keyword.has_key?(opts, key), do: opts, else: Keyword.put(opts, key, value)
-  end
-
-  # Read-only by default (fs read tools + grep/glob). `--write` adds the
-  # mutating coding tools (write_file/edit_file/bash), which are `sensitive`
-  # and denied under the default policy — so it also installs an allow-all
-  # authorizer to actually let them run in this unattended headless flow.
-  #
-  # The benchmark profile forces write-mode semantics (no human to ask; the
-  # task container is the blast radius) and drops skills entirely: task
-  # attempts must be independent, so no cross-run skill loop.
-  defp actions_for(_write?, %BenchmarkProfile{active?: true}),
-    do: Raxol.Agent.Actions.Fs.all() ++ Raxol.Agent.Actions.Code.all()
-
-  defp actions_for(false, _profile),
-    do:
-      Raxol.Agent.Actions.Fs.all() ++
-        Raxol.Agent.Actions.Code.read_only() ++ Raxol.Agent.Skills.enabled_actions()
-
-  defp actions_for(true, _profile),
-    do:
-      Raxol.Agent.Actions.Fs.all() ++
-        Raxol.Agent.Actions.Code.all() ++ Raxol.Agent.Skills.enabled_actions()
-
-  defp context_for(_write?, %BenchmarkProfile{active?: true}),
-    do: [context: %{tool_authorizer: Raxol.Agent.ToolPolicy.allow_all()}]
-
-  defp context_for(false, _profile), do: maybe_skills_context(%{})
-
-  defp context_for(true, _profile),
-    do: maybe_skills_context(%{tool_authorizer: Raxol.Agent.ToolPolicy.allow_all()})
-
-  # Add the configured skills store under context[:skills]; keep the empty-context
-  # `[]` shape when nothing (skills or authorizer) needs to be passed.
-  defp maybe_skills_context(base) do
-    base =
-      case Raxol.Agent.Skills.default_context() do
-        nil -> base
-        skills -> Map.put(base, :skills, skills)
-      end
-
-    if map_size(base) == 0, do: [], else: [context: base]
-  end
-
-  defp maybe_put(kw, _key, nil), do: kw
-  defp maybe_put(kw, key, value), do: Keyword.put(kw, key, value)
-
-  # -- Event consumption: contract events in, stdout/stderr out --------------
-
-  defp consume(session_id, runner, timeout_ms, state) do
-    receive do
-      {:session_event, ^session_id, %Contract.Event{} = event} ->
-        IO.write(:stderr, Contract.encode_line(event))
-
-        state =
-          event
-          |> render_stdout(state)
-          |> track(event)
-
-        case event do
-          %{type: :turn_completed, payload: %{final: true}} ->
-            Task.await(runner, 5_000)
-            if state.wrote_stdout, do: IO.write("\n")
-            finish(state, 0, :completed)
-
-          %{type: :error} ->
-            Task.await(runner, 5_000)
-            finish(state, 1, :error)
-
-          _ ->
-            check_budget_then_continue(session_id, runner, timeout_ms, state)
-        end
-
-      {:os_signal, :sigterm} ->
-        IO.puts(
-          :stderr,
-          ~s({"type":"error","payload":{"reason":"terminated"}})
-        )
-
-        Task.shutdown(runner, :brutal_kill)
-        finish(state, 143, :terminated)
-    after
-      timeout_ms ->
-        IO.puts(:stderr, ~s({"type":"error","payload":{"reason":"timeout"}}))
-        Task.shutdown(runner, :brutal_kill)
-        finish(state, 2, :timeout)
-    end
-  end
-
-  defp check_budget_then_continue(session_id, runner, timeout_ms, state) do
-    case BenchmarkProfile.budget_status(state.profile, state.turns, state.usage) do
-      :ok ->
-        consume(session_id, runner, timeout_ms, state)
-
-      {:exceeded, cap} ->
-        IO.puts(
-          :stderr,
-          ~s({"type":"error","payload":{"reason":"budget_exhausted","cap":"#{cap}"}})
-        )
-
-        Task.shutdown(runner, :brutal_kill)
-        finish(state, 2, :budget_exhausted)
-    end
-  end
-
-  # Accumulate turn/usage totals and (when a trajectory is requested) the
-  # event list itself.
-  defp track(state, %{type: :turn_completed, payload: payload} = event) do
-    state
-    |> Map.update!(:turns, &(&1 + 1))
-    |> Map.update!(:usage, &BenchmarkProfile.add_usage(&1, Map.get(payload, :usage, %{})))
-    |> record(event)
-  end
-
-  defp track(state, event), do: record(state, event)
-
-  defp record(%{events: nil} = state, _event), do: state
-  defp record(state, event), do: Map.update!(state, :events, &[event | &1])
-
-  # Every exit path flushes the trajectory (when configured) and returns the
-  # status for `exit({:shutdown, status})`.
-  defp finish(state, exit_code, reason) do
-    case state.profile.trajectory_path do
-      nil ->
-        :ok
-
-      path ->
-        trajectory =
-          Trajectory.build(Enum.reverse(state.events || []), %{
-            prompt: state.prompt,
-            backend: state.backend,
-            model: state.profile.model,
-            profile: state.profile,
-            turns: state.turns,
-            usage: state.usage,
-            exit_code: exit_code,
-            reason: reason
-          })
-
-        Trajectory.write(path, trajectory)
-    end
-
-    exit_code
-  end
-
-  # stdout carries the ANSWER only. Stream deltas as they arrive; if the
-  # run produced no deltas (non-streaming react loop), print the final
-  # message content once.
-  defp render_stdout(%{type: :item_delta, payload: %{chunk: chunk}}, state) do
-    IO.write(chunk)
-    %{state | wrote_stdout: true}
-  end
-
-  defp render_stdout(
-         %{
-           type: :item_completed,
-           payload: %{item_type: :message, content: content}
-         },
-         %{wrote_stdout: false} = state
-       ) do
-    IO.write(content)
-    %{state | wrote_stdout: true}
-  end
-
-  defp render_stdout(_event, state), do: state
 end
